backend/requirements.txt
asgiref==3.4.1
colorama==0.4.4
commonmark==0.9.1
Django==3.2.7
django-appconf==1.0.5
django-cors-headers==3.8.0
django-imagekit==4.0.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
Faker==8.14.0
filetype==1.0.8
mixer==7.1.2
mysqlclient==2.0.3
pilkit==2.0
Pillow==8.3.2
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.1
rich==10.11.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
